trurl: abort on a bad URL
Slight change of behavior: if a URL is provided that cannot be parsed,
it will now abort the operation for that URL and move on to the next URL
instead. It makes for more consistent and dependable behavior.

Adding --verify to the mix will make trurl abort and exit with an error
for the first non-parseable URL.

This probably replaces #84 and also fixes #76.

Fixes #76
Fixes #84
Closes #91