handle accept-space in older libcurl versions
If CURLU_ALLOW_SPACE is false, and --allow-space is requested,
we emit a warning and disable its behavior as it is unsupported.

fix segfault in macro implementation of curl_url_strerror

The use of fprintf and curl_easy_strerror is erronous as uses of
curl_url_strerror expect a const char * return, resulting in crashing.