trurl: make '{scheme}\{user}' work
Before this patch, the sequence `\{' was handled by printing a literal
backslash followed by a literal {.

This was a bit awkward since you are supposed to be using `{{' to escape
a { character, not \. Also because the backslash was still being
printed.

With this patch `\{' is now handled by only printing a backslash, and
then handling the next character, instead of printing it
indiscrimintately.

So, now,   trurl -g '{scheme}\{user}' gopher://emanuele6@locahost/
outputs   gopher\emanuele6   instead of   gopher\{user}   .

The previous commit that added support for `\\' makes this change less
necessary since the user should have used   {scheme}\\{user}   instead
of using a single backslash, but I still think this behaviour makes more
sense, and may prevent future bugs.