Respect CFLAGS, CPPFLAGS in Makefile, fix LDFLAGS order
The implicit rule handled this correctly but a7f9ca9c42c2cf29cc84becb38b56082898868d4
did away with that.

Add CPPFLAGS (for the C preprocessor) and CFLAGS explicitly, and then shift
LDFLAGS earlier to be before the object list, which is needed for e.g. -Wl,--as-needed
to work correctly.

Fixes: a7f9ca9c42c2cf29cc84becb38b56082898868d4