test: print failed tests more nicely.
Instead of printing '{rawstring}' or '{stringifiedpythonobject}', print
the python representation of the value using f"{foo!r}".

This means that objects are not surrounded by quotes, and that strings
with newlines (basically all of them), display newlines as \n and are
printed on a single line.
Also numbers are not surrounded by quotes, so if someone enters
"return_code":"7" instead of "return_code":7 in tests.json by mistake,
they see expected:'7' got:7 in the report instead of
expected:'7' got:'7'. :)

Overall the output is nicer to look at and less ambiguous.