--redirect: don't silently ignore failed redirections.
If a redirection fails, report the failure to stderr, and go to the next
URL, or abort if running with --verify. Instead of silently ignoring the
failure and carrying on.

This patch also renames urlfromstring() to seturl(), and implements
--redirect with it so that curl_url_set() is called with the same flags
used to parse base URLs.

Which means the following command will now work:

  $ trurl --accept-space --redirect 'x://foo/a b' curl.se
  x://foo/a%20b