trurl.1: various improvements
Remove trailing spaces from some lines, and fix some typos.

Capitalise ASCII, and URL where it makes sense.

Clarify that URL arguments can be specified without --url, and mention
-- support.

Mention explicitly (not just in the examples), that if no base URL is
provided, trurl will create a URL using the specified components.

Clarify that --redirect can be used even without --url as long as a base
URL is specified.

Line limit when reading --url-file files is not 4095, it is 4094 bytes
(4096 - '\n' - '\0'). Also mention that trurl skips empty lines, and
ignores '\r' at the end of the line.

Co-Authored-By: Daniel Stenberg