trurl: restructure JSON, add --default-port, --keep-port, --punycode
This patch restructures the JSON output with the following format:

  {
    "url": // same value as {url}
    "parts": {
      ..., // urldecoded URL parts except the URL
    },
    "params": [
       ...
    ],
    // other stuff
  }

I decided to remove raw_port for now, since now I think it is less
useful than it used to be.

This patch also reworks default output and -g.
{raw:} has been removed in favour of {default:}.
{default:} and {puny:} are "get modifiers" like {:component}, and
may be specified in any order front of a component to modify the
curl_url_get() flags used by trurl to get that URL.

{default:} makes trurl use CURLU_DEFAULT_PORT, use the libcurl-supported
scheme's default port if one is not specified, and {puny:} makes trurl
use CURLU_PUNYCODE  e.g.

* {default:port} will get 80 for http://example.org/
* {puny:host} will get xn--p-8fa.example.org from imap://pè.example.org

{default:}, {puny:}, and {:part} may be specified at the same time in
which case multiple additional flags are passed to curl_url_get().

  $ trurl -g '{puny:default:url}' imap://pè.example.org
  imap://xn--p-8fa.example.org:143/

If --default-port is used, {default:} is implied for every {} in the
--get; and, if --puny is implied.

Default output now behaves exactly like  -g '{url}'.
(note: previously  -g '{url}'  behaved like the new {default:url}, it
       always included, )

This allows more fine tuning of `-g` the output, and default URL output
since it was previously not possible to:

* get the URL with explicit default port in the --json output
* get the URL without default port in a --get format
* get the punycode encoded URL without explicit default port
* get the punycode encoded URL with or without explict port in --json

By default, trurl removes redundant explict ports for libcurl-supported
schemes, you can use --keep-port to inhibit that behaviour.

I fixed a bug that made  -g {raw:port}  (or  -g {port}) never able to
print the default port for libcurl-supported scheme even if the port was
explictly specified. (this bug was unique to {raw:port}, raw_port in the
JSON worked fine.)

I also fixed some inconsistencies in the code, e.g. get() had a output
stream variable that was not being propagated to showqkey() that was
always outputting to stdout, and failed gets were printing warnings
using  printf(stderr, PROGNAME ...)  instead of warnf().

This is only an initial draft, so I have not updates test expectations
and the manual yet.