trurl: support ?= for set, to only do it if not already set
Idea-by: @emanuele6 in #296

Due to missing support in the libcurl URL API, this is not really
working for the scheme component right now, as that will always be
considered set once a URL has been parsed. The scheme guessing the
parser does sets it internally and there is currently no way provided by
the API to figure that out.