trurl: make must: prefix for gets return error for missing part
Normally a missing component just causes a blank output. This currently
does not work "scheme" (it defaults to the "guessed" scheme).

"must:" deos not return error if query/fragment was there but empty

(this latter part requires libcurl 8.8.0 or later. The feature is named
'get-empty' in the version output.)

Added tests

Closes #336